Outsource YRedisUser in additonal file.
