+++ src/infra/y-redis/ws.service.ts
@@ -15,6 +15,7 @@ import { YRedisClient } from './y-redis.client.js';
 import { computeRedisRoomStreamName, isSmallerRedisId } from './helper.js';
 import * as protocol from './protocol.js';
 import { Subscriber } from './subscriber.service.js';
+import { YRedisUser } from './y-redis-user.js';
 
 /**
  * how to sync
@@ -27,7 +28,7 @@ import { Subscriber } from './subscriber.service.js';
  *     subscriber.ensureId(redisDocLastId)
  *   }
  */
-
+/* 
 export class YWebsocketServer {
 	public constructor(
 		public readonly app: uws.TemplatedApp,
@@ -40,39 +41,7 @@ export class YWebsocketServer {
 		await this.client.destroy();
 	}
 }
-
-let _idCnt = 0;
-
-export class YRedisUser {
-	public subs: Set<string>;
-	public id: number;
-	public awarenessId: number | null;
-	public awarenessLastClock: number;
-	public isClosed: boolean;
-	public initialRedisSubId: string;
-
-	public constructor(
-		public readonly room: string | null,
-		public readonly hasWriteAccess: boolean,
-		/**
-		 * Identifies the User globally.
-		 * Note that several clients can have the same userid (e.g. if a user opened several browser
-		 * windows)
-		 */
-		public readonly userid: string | null,
-		public readonly error: Partial<CloseEvent> | null = null,
-	) {
-		this.initialRedisSubId = '0';
-		this.subs = new Set();
-		/**
-		 * This is just an identifier to keep track of the user for logging purposes.
-		 */
-		this.id = _idCnt++; // TODO
-		this.awarenessId = null;
-		this.awarenessLastClock = 0;
-		this.isClosed = false;
-	}
-}
+*/
 
 export const upgradeCallback = async (
 	res: uws.HttpResponse,
